# nicloud/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PRIVATE CLOUD PLATFORM",
        "title": "NILCOUD",
        "termsOfService": "https://github.com/niklaus-code/nicloud",
        "contact": {
            "name": "NIKLAUS",
            "url": "https://github.com/niklaus-code/nicloud",
            "email": "1309584951@qq.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录接口1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "usernmae",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "passwd",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.Vms_users"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "users.Vms_users": {
            "type": "object",
            "required": [
                "Passwd",
                "Username"
            ],
            "properties": {
                "Mobile": {
                    "type": "string"
                },
                "Passwd": {
                    "type": "string"
                },
                "Role": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "Username": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}